🎨 style(form.scss): update styles for .fa-chevron-down and .fa-chevron-left The styles for .fa-chevron-down and .fa-chevron-left have been updated to include the color $gray. This ensures consistency in the color scheme across the application. Additionally, the .custom--form--dropdown--container class has been removed as it is no longer needed. The .custom--form--field class now has a border of 2.5px solid $mid-gray for a more visually appealing form area.
🐛 fix(dropdown_controller.js): fix syntax errors and add missing methods
✨ feat(dropdown_controller.js): add functionality to open and close dropdown, render rooms
Fixed syntax errors in the code and added missing methods. The connect method now sets the active room in the dropdown. The open method opens the dropdown and renders the list of rooms. The close method closes the dropdown. The #renderRooms method renders the list of inactive rooms in the dropdown.

🔧 fix(belongings/_form.html.erb): update dropdown behavior and remove unused code
🎨 style(belongings/_form.html.erb): improve dropdown styling and remove unnecessary classes
The dropdown behavior in the form has been updated to use a data attribute and a corresponding JavaScript controller. The code for the dropdown options has been removed as it is now generated dynamically by the controller. The styling of the dropdown has been improved by removing unnecessary classes and adding a chevron icon for better visual indication.