🎨 style(application.scss): update padding and flex property in .page-container 🎨 style(footer.scss): update padding in .footer and .footer a 🎨 style(layout.scss): add display and flex-direction properties in .core-page-layout, add overflow, flex, and display properties in .user-content-page
The padding in .page-container is updated to have 15px top and bottom padding, and the flex property is added with a value of 1.

The padding in .footer and .footer a is updated to have 0px left and right padding.

In .core-page-layout, the display property is added with a value of flex, and the flex-direction property is set to column. The height property is removed, and the justify-content property is set to space-between.

In .user-content-page, the overflow property is set to scroll, the flex property is set to 1, and the display property is set to flex with

🎨 style(_flats_index.scss): fix indentation and add spaces around border property
🐛 fix(_flats_index.scss): fix indentation and add spaces around border property
🐛 fix(_flats_index.scss): fix indentation and add spaces around border property
🎨 style(_flats_index.scss): fix indentation and add spaces around border property
🎨 style(_flats_index.scss): fix indentation and add spaces around border property
🎨 style(_flats_index.scss): fix indentation and add spaces around border property
🎨 style(_flats_index.scss): fix indentation and add spaces around border property
🎨 style(_flats_index.scss): fix indentation and add spaces around border property
🎨 style(_flats_index.scss): fix indentation and add spaces around border property
🎨 style(_flats_index.scss): fix indentation and add spaces around border property
🎨 style(_fl

🐛 fix(layout_controller.js): fix body height calculation in connect method
The body height calculation in the connect method of the layout controller has been fixed. Instead of setting the margin top, the body height is now calculated using the navbar height to ensure that the content fits within the available viewport height.

🎨 style(_navbar.html.erb): remove fixed-top class from navbar
The fixed-top class has been removed from the navbar element in order to improve the layout and styling of the navigation bar.

🔀 refactor(_navbar.html.erb): reorganize and update navigation menu
🎨 style(_navbar.html.erb): remove unnecessary closing div tag
The navigation menu in the shared _navbar.html.erb file has been reorganized and updated. The changes include:
- Moving the user profile image and email to a separate div for better styling.
- Reordering the menu items and adding arrows for better navigation.
- Updating the logout link to use the data-turbo-method attribute for improved security.
- Removing an unnecessary closing div tag.

🎨 style(flats/index.html.erb): remove unnecessary top margin from welcome-flat-index class
🎨 style(layouts/application.html.erb): add user-content-page class to the main content container
🎨 style(pages/home.html.erb): change logout link text to Signin
The top margin of the welcome-flat-index class has been removed to improve the visual appearance of the page. The user-content-page class has been added to the main content container in the application layout to differentiate it from the landing page. The logout link text on the home page has been changed to Signin to provide clearer navigation options.