🎨 style(footer.scss): add padding-bottom to footer component 🎨 style(footer.scss): adjust padding in footer links 🎨 style(footer.scss): add padding to footer copyright 🎨 style(navbar.scss): remove height from navbar 🎨 style(navbar.scss): adjust padding in navbar 🎨 style(navbar.scss): remove margin and padding from menu-icon 🎨 style(navbar.scss): adjust padding and display in navbar 🎨 style(navbar.scss): remove margin and padding from navbar-toggler 🎨 style(navbar.scss): adjust margin in profile-pic-nav 🎨 style(navbar.scss): add flex properties to logo-container 🎨 style(navbar.scss): adjust width in navbar-logo
The changes in the footer.scss file include adding a padding-bottom to the footer component, adjusting the padding in the footer links, and adding padding to the footer copyright.

In the navbar

🔧 fix(navbar.html.erb): refactor navbar structure and remove unnecessary divs
The navbar structure has been refactored to remove unnecessary divs and improve code readability. The changes include removing the outer div with the class "d-flex", removing the div with the class "profile-menu", and reorganizing the elements within the navbar.